views 분리
__init__
answer_views
question_views
1번 방법
