+++ .gitignore
@@ -1,0 +1,4 @@
+
+*.xml
+*.iml
+*.pyc
