+++ mysite/pybo/views/__init__.py
@@ -1,0 +1,3 @@
+from .base_views import *
+from  .question_views import *
+from .answer_views import *
